session_db: adapt to OCA standards
