+++ session_db/__init__.py
@@ -1,1 +1,1 @@
-import models
+from . import pg_session_store
