+++ session_db/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Store sessions in a database instead of the filesystem.
