+++ session_db/readme/ROADMAP.rst
@@ -1,0 +1,1 @@
+This module does not work with multi-threaded workers, so it requires workers > 0.
