+++ session_db/readme/USAGE.rst
@@ -1,0 +1,7 @@
+Set this module in the server wide modules.
+
+Set a ``SESSION_DB_URI`` environment variable as a full postgresql connection string,
+like ``postgres://user:passwd@server/db`` or ``db``.
+
+It is recommended to use a dedicated database for this module, and possibly a dedicated
+postgres user for additional security.
